Initial Code Commit
Basic routing and note object. Note object has an id, title, and body. Get, post, delete, and put are implemented temporarily using an array to hold the data.

This is what I've been working on the past 8 days. Everyday I try to do a little more, so I thought I'd track it on here.