File Storage Implementation, Array of Structs No Longer Used
Previously, an array of structs was used to represent the data. It was stored in ram, and not as json files.

In this push, all routing functions no longer use the array to access memory, but the actual data files themselves.

Also, the original post method had a bug (I believe this was in the last push) in which ID's were not unique. They are now unique, and are properly integrated with file handling.