Initial Angular Setup, Reorganized Folders
Couldn't sleep so I did a little bit of reorganizing. The previous version of this project was moved to a folder called "server" and the unnecessary "server.go" file was removed. This was used during early development as a reference.

Another folder was created, "neuron-frontend" (probably will be renamed at a later date) and will hold the frontend component during this testing phase, if you will.

I still have to learn a lot about Angular. Honestly this is my first time using a front-end framework, so there's gonna be a curve here. I'd imagine the frontend part will take a little bit longer than may be expected, but I'll continue to pick at it and add components of the frontend as I go.