Added Error Responses
In previous versions, there was checks for errors and the function would simply return. Now, there is a basic response before the function return to potential errors.

In the future I would like to do some research as to what the optimal status responses would be. This is a good first step though, and a much needed change.